rename the file again
