+++ regionmask/defined_regions/__init__.py
@@ -1,7 +1,7 @@
 # flake8: noqa
 
 from ._ar6 import ar6
-from ._ar6_pre_rev import _ar6_pre_revisions
+from ._ar6_pre_revisions import _ar6_pre_revisions
 from .giorgi import giorgi
 from .natural_earth import _maybe_get_column, _obtain_ne, natural_earth
 from .srex import srex
